refactored draggable courses into its own component
