+++ src/pages/Home.tsx
@@ -1,7 +1,6 @@
 import AddCourseModal from "../components/Course/AddCourse/AddCourseModal";
 import React from "react";
 import { Section } from "../common/types";
-import CourseDropdown from "../components/Course/CourseDropdown";
 import SidebarContainer from "../components/Layout/SidebarContainer";
 import { Box } from "@mui/material";
 import Schedule from "../components/Schedule/Schedule";
@@ -9,22 +8,7 @@ import usePlans from "../hooks/usePlans";
 import AddPlanModal from "../components/Plan/AddPlanModal";
 import SelectPlan from "../components/Plan/SelectPlan";
 import PlanInfoDisplay from "../components/Plan/PlanInfoDisplay";
-import {
-  DndContext,
-  closestCenter,
-  KeyboardSensor,
-  PointerSensor,
-  useSensor,
-  useSensors,
-} from "@dnd-kit/core";
-import {
-  SortableContext,
-  verticalListSortingStrategy,
-  sortableKeyboardCoordinates,
-} from "@dnd-kit/sortable";
-import { SortableItem } from "../components/Course/SortableItem";
-import { arrayMove } from "../common/utils";
-import { UniqueIdentifier } from "@dnd-kit/core";
+import DraggableCourses from "../components/Course/DraggableCourses";
 
 function Home() {
   const {
@@ -46,29 +30,6 @@ function Home() {
     setPlan(emptyPlan);
   }
 
-  function handleDragEnd(event: {
-    active: { id: UniqueIdentifier };
-    over: { id: UniqueIdentifier } | null;
-  }) {
-    const { active, over } = event;
-
-    if (over?.id && active.id !== over.id) {
-      const oldIndex = plan.courses.findIndex(
-        (course) => course.classId === active.id.toString()
-      );
-      const newIndex = plan.courses.findIndex(
-        (course) => course.classId === over.id.toString()
-      );
-
-      if (oldIndex !== -1 && newIndex !== -1) {
-        setPlan({
-          ...plan,
-          courses: arrayMove(plan.courses, oldIndex, newIndex),
-        });
-      }
-    }
-  }
-
   const [hoverSection, setHoverSection] = React.useState<Section[]>([]);
 
   return (
@@ -79,47 +40,14 @@ function Home() {
           handleRemovePlan={handleRemovePlan}
           setPlanName={setPlanName}
         />
-        <Box sx={{ width: "100%" }}>
-          <DndContext
-            sensors={useSensors(
-              useSensor(PointerSensor),
-              useSensor(KeyboardSensor, {
-                coordinateGetter: sortableKeyboardCoordinates,
-              })
-            )}
-            collisionDetection={closestCenter}
-            onDragEnd={handleDragEnd}
-          >
-            <SortableContext
-              items={plan.courses.map((course) => course.classId)}
-              strategy={verticalListSortingStrategy}
-            >
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  alignItems: "center",
-                }}
-              >
-                {plan.courses.map((course) => (
-                  <SortableItem key={course.classId} id={course.classId}>
-                    {(dragHandleProps) => (
-                      <CourseDropdown
-                        setHoverSection={setHoverSection}
-                        plan={plan}
-                        course={course}
-                        removeCourse={removeCourse}
-                        addSection={addSection}
-                        removeSection={removeSection}
-                        dragHandleProps={dragHandleProps}
-                      />
-                    )}
-                  </SortableItem>
-                ))}
-              </Box>
-            </SortableContext>
-          </DndContext>
-        </Box>
+        <DraggableCourses
+          plan={plan}
+          setPlan={setPlan}
+          setHoverSection={setHoverSection}
+          removeCourse={removeCourse}
+          addSection={addSection}
+          removeSection={removeSection}
+        ></DraggableCourses>
         <Box
           sx={{
             display: "flex",
